DMT implemented in Dido using a Path Constraint, section 3.5 of text
Implementation of problem DMT in section 3.5 of Michael Ross' text "A
Primer on Pontryagin's Principle in Optimal Control". By implementing
the box-bounds on the control as a path constraint, we get access to the
dual variable that defines the switching structure of the optimal
control, which comes from the complementary condition part of the
necessary conditions given by KKT when solving the Hamiltonian
Minimization Condition.